fix: refactor DateTimePicker
separate TimePicker and DatePicker as standalone components

Fixes #238
