+++ src/components/DateTimePicker/index.js
@@ -1,55 +1,27 @@
-import React, { forwardRef, useEffect, useState } from 'react'
-import { createPortal } from 'react-dom'
-import { bool, func, number, object, oneOf, oneOfType, string } from 'prop-types'
+import React, { cloneElement, forwardRef, useEffect, useState } from 'react'
+import { func, node, number, object, oneOfType, string } from 'prop-types'
 
-import { COMPONENT_TYPE, SIZES_TYPE } from '../../utils'
+import { SIZES_TYPE } from '../../utils'
+import { getDate } from '../DateTimePicker/utils'
+import { DatePicker } from '../DatePicker'
+import { TimePicker } from '../TimePicker'
 
-import { CustomInput } from './CustomInput'
-import { getDate } from './utils'
 import * as S from './styles'
 
-const DEFAULT_INTERVAL = 15
-const DEFAULT_DATE_FORMAT = 'dd/MM/yyyy'
-const DEFAULT_TIME_FORMAT = 'HH:mm'
-
-const PopperPortal = ({ children }) =>
-  typeof window !== 'undefined'
-    ? createPortal(<S.Popper>{children}</S.Popper>, document.querySelector('body'))
-    : null
-
 export const DateTimePicker = forwardRef(
-  (
-    {
-      value = new Date(),
-      dateIcon,
-      dateIconPlacement = 'left',
-      datePickerOnly,
-      datePickerProps = {},
-      onChange,
-      size = 'lg',
-      timeIcon,
-      timeIconPlacement = 'left',
-      timePickerOnly,
-      timePickerProps = {},
-      ...rest
-    },
-    ref
-  ) => {
-    // Set defaults
-    const datePickerDateFormat = datePickerProps.dateFormat || DEFAULT_DATE_FORMAT
-    const timePickerDateFormat = timePickerProps.dateFormat || DEFAULT_TIME_FORMAT
-    const timeIntervals = timePickerProps.timeIntervals || DEFAULT_INTERVAL
+  ({ children, onChange, size = 'lg', value = new Date() }, ref) => {
+    const [date, setDate] = useState(value)
 
-    const [focused, setFocused] = useState(
-      (datePickerProps.autoFocus && 'date') || (timePickerProps.autoFocus && 'time') || null
-    )
-    const [date, setDate] = useState(null)
+    const handleChange = newDate => {
+      setDate(newDate)
+      onChange && onChange(new Date(newDate))
+    }
 
-    const formatDate = date => getDate(date, timeIntervals)
+    const formatDate = date => getDate(date, 15)
 
     // format date at component mount
     useEffect(() => {
-      onChange && onChange(formatDate(value))
+      onChange && handleChange(formatDate(value))
       //eslint-disable-next-line
     }, [])
 
@@ -57,100 +29,29 @@ export const DateTimePicker = forwardRef(
     useEffect(() => {
       const formattedDate = formatDate(value)
       if (value - formattedDate !== 0 && onChange) {
-        onChange(formattedDate)
+        handleChange(formattedDate)
       }
       setDate(formattedDate)
       //eslint-disable-next-line
     }, [value])
 
-    const handleFocus = (kind, e) => {
-      let onDatePickerFocus = datePickerProps.onFocus
-      let onTimePickerFocus = timePickerProps.onFocus
-      setFocused(kind)
-      kind === 'date' && onDatePickerFocus && onDatePickerFocus(e)
-      kind === 'time' && onTimePickerFocus && onTimePickerFocus(e)
-    }
-
-    const handleBlur = (kind, e) => {
-      let onDatePickerBlur = datePickerProps.onBlur
-      let onTimePickerBlur = timePickerProps.onBlur
-      setFocused(null)
-      kind === 'date' && onDatePickerBlur && onDatePickerBlur(e)
-      kind === 'time' && onTimePickerBlur && onTimePickerBlur(e)
-    }
-
-    const handleChange = (newDate, kind) => {
-      if (!newDate) return
-      const date = formatDate(newDate)
-
-      if (kind === 'date') {
-        newDate = newDate.setHours(date.getHours(), date.getMinutes())
-      }
-      if (kind === 'time') {
-        newDate = newDate.setFullYear(date.getFullYear(), date.getMonth(), date.getDate())
-      }
-      setDate(newDate)
-      onChange && onChange(new Date(newDate))
-    }
-
     return (
-      <S.DateTimePicker
-        data-testid="dateTimePicker"
-        datePickerOnly={datePickerOnly}
-        timePickerOnly={timePickerOnly}
-        {...rest}
-      >
-        {!timePickerOnly && (
-          <S.DatePicker
-            calendarClassName="date-picker-popper"
-            className="date-picker"
-            customInput={
-              <CustomInput
-                focused={focused}
-                handleBlur={e => handleBlur('date', e)}
-                handleFocus={e => handleFocus('date', e)}
-                icon={dateIcon}
-                iconPlacement={dateIconPlacement}
-                inputRef={timePickerOnly ? null : ref}
-                ref={ref}
-                size={size}
-              />
-            }
-            popperContainer={PopperPortal}
-            {...datePickerProps}
-            dateFormat={datePickerDateFormat}
-            onChange={date => handleChange(date, 'date')}
-            selected={date}
-            size={size}
-            {...rest}
-          />
-        )}
-        {!datePickerOnly && (
-          <S.TimePicker
-            calendarClassName="time-picker-popper"
-            className="time-picker"
-            customInput={
-              <CustomInput
-                focused={focused}
-                handleBlur={e => handleBlur('date', e)}
-                handleFocus={e => handleFocus('date', e)}
-                icon={timeIcon}
-                iconPlacement={timeIconPlacement}
-                inputRef={timePickerOnly ? ref : null}
-                size={size}
-              />
-            }
-            popperContainer={PopperPortal}
-            {...timePickerProps}
-            dateFormat={timePickerDateFormat}
-            onChange={time => handleChange(time, 'time')}
-            selected={date}
-            showTimeSelect
-            showTimeSelectOnly
-            size={size}
-            timeIntervals={timeIntervals}
-            {...rest}
-          />
+      <S.DateTimePicker data-testid="dateTimePicker" size={size}>
+        {children &&
+          children.map((child, i) =>
+            cloneElement(child, {
+              key: i,
+              onChange: handleChange,
+              // give ref only to the first child
+              inputRef: i < 1 && ref,
+              value: date
+            })
+          )}
+        {!children && (
+          <>
+            <DatePicker inputRef={ref} onChange={handleChange} size={size} value={value} />
+            <TimePicker onChange={handleChange} size={size} value={value} />
+          </>
         )}
       </S.DateTimePicker>
     )
@@ -160,15 +61,8 @@ export const DateTimePicker = forwardRef(
 DateTimePicker.displayName = 'DateTimePicker'
 
 DateTimePicker.propTypes = {
-  dateIcon: COMPONENT_TYPE,
-  dateIconPlacement: oneOf('right', 'left'),
-  datePickerOnly: bool,
-  datePickerProps: object,
+  children: node,
   onChange: func,
   size: SIZES_TYPE,
-  timeIcon: COMPONENT_TYPE,
-  timeIconPlacement: oneOf('right', 'left'),
-  timePickerOnly: bool,
-  timePickerProps: object,
   value: oneOfType([number, object, string]).isRequired
 }
